Remove deprecated Dummy payment gateway plugins
Drop DeprecatedDummyGatewayPlugin (mirumee.payments.dummy) and
DeprecatedDummyCreditCardGatewayPlugin (mirumee.payments.dummy_credit_card)
from BUILTIN_PLUGINS, so they are no longer part of the shipped plugin set.

The legacy plugin-based Payment flow (checkoutPaymentCreate,
payment{Capture,Refund,Void}, order{Capture,Refund,Void}, checkoutComplete
with a payment) is still part of the product, and these were the only
gateways its ~120 tests could run against. The implementations are therefore
kept as explicit test doubles under saleor/plugins/tests/gateways/, renamed
to DummyGatewayPlugin / DummyCreditCardGatewayPlugin. PLUGIN_IDs are
unchanged, and each gateway's module and plugin class are now collapsed into
a single module.

Test references and patch targets updated accordingly.
