+++ saleor/graphql/checkout/tests/mutations/test_checkout_complete_with_payment.py
@@ -10,6 +10,8 @@ from django.db.models.aggregates import Sum
 from django.test import override_settings
 from django.utils import timezone
 
+from saleor.plugins.tests.gateways.dummy_credit_card import TOKEN_VALIDATION_MAPPING
+
 from .....account.models import Address
 from .....checkout import calculations
 from .....checkout.delivery_context import (
@@ -33,7 +35,6 @@ from .....order.delivery_context import PRIVATE_META_APP_SHIPPING_ID
 from .....order.models import Fulfillment, Order
 from .....payment import ChargeStatus, PaymentError, TransactionKind
 from .....payment.error_codes import PaymentErrorCode
-from .....payment.gateways.dummy_credit_card import TOKEN_VALIDATION_MAPPING
 from .....payment.interface import GatewayResponse
 from .....payment.model_helpers import get_subtotal
 from .....plugins.manager import PluginsManager, get_plugins_manager
@@ -2981,8 +2982,8 @@ def test_checkout_complete_checkout_without_lines(
 @pytest.mark.integration
 @pytest.mark.parametrize(("token", "error"), list(TOKEN_VALIDATION_MAPPING.items()))
 @patch(
-    "saleor.payment.gateways.dummy_credit_card.plugin."
-    "DeprecatedDummyCreditCardGatewayPlugin.DEFAULT_ACTIVE",
+    "saleor.plugins.tests.gateways.dummy_credit_card."
+    "DummyCreditCardGatewayPlugin.DEFAULT_ACTIVE",
     True,
 )
 def test_checkout_complete_error_in_gateway_response_for_dummy_credit_card(
@@ -5754,7 +5755,7 @@ def test_checkout_complete_empty_product_translation(
 @override_settings(
     PLUGINS=[
         "saleor.plugins.webhook.plugin.WebhookPlugin",
-        "saleor.payment.gateways.dummy.plugin.DeprecatedDummyGatewayPlugin",
+        "saleor.plugins.tests.gateways.dummy.DummyGatewayPlugin",
     ]
 )
 @patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
@@ -5857,7 +5858,7 @@ def test_checkout_complete_with_external_shipping(
 @override_settings(
     PLUGINS=[
         "saleor.plugins.webhook.plugin.WebhookPlugin",
-        "saleor.payment.gateways.dummy.plugin.DeprecatedDummyGatewayPlugin",
+        "saleor.plugins.tests.gateways.dummy.DummyGatewayPlugin",
     ]
 )
 @patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
