+++ saleor/graphql/invoice/tests/test_invoice_request_delete.py
@@ -24,7 +24,7 @@ INVOICE_REQUEST_DELETE_MUTATION = """
 @pytest.fixture(autouse=True)
 def setup_dummy_gateways(settings):
     settings.PLUGINS = [
-        "saleor.payment.gateways.dummy.plugin.DeprecatedDummyGatewayPlugin",
+        "saleor.plugins.tests.gateways.dummy.DummyGatewayPlugin",
     ]
     return settings
