+++ saleor/tests/e2e/checkout/shipping/test_use_external_shipping_methods_in_checkout.py
@@ -30,7 +30,7 @@ def test_use_external_shipping_methods_in_checkout_core_1652(
     # Before
     settings.PLUGINS = [
         "saleor.plugins.webhook.plugin.WebhookPlugin",
-        "saleor.payment.gateways.dummy.plugin.DeprecatedDummyGatewayPlugin",
+        "saleor.plugins.tests.gateways.dummy.DummyGatewayPlugin",
     ]
     permissions = [
         permission_manage_products,
